feat: :sparkles: 맥주목록 > BeerSearchResultEmpty 추가, 리팩토링
