--- src/pages/beers/components/BeerListPageHeader/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BeerListPageHeader';
