+++ src/pages/beers/components/SearchBox/SearchBox.tsx
