+++ src/pages/beers/components/SearchBox/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SearchBox';
